+++ app/kuma-prometheus-sd/cmd/root_helper_test.go
@@ -1,0 +1,6 @@
+package cmd
+
+var (
+	// make `newRootCmd` available to `cmd_test`
+	NewRootCmd = newRootCmd
+)