+++ app/kuma-prometheus-sd/cmd/root_test.go
@@ -1,0 +1,20 @@
+package cmd_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	. "github.com/Kong/kuma/app/kuma-prometheus-sd/cmd"
+)
+
+var _ = Describe("root", func() {
+	It("should be possible to run `kuma-prometheus-sd` without a sub-command", func() {
+		// given
+		cmd := NewRootCmd()
+		cmd.SetArgs([]string{})
+		// when
+		err := cmd.Execute()
+		// then
+		Expect(err).ToNot(HaveOccurred())
+	})
+})